requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
